# Optimize `AsyncLazy`