Master:

|    Method | IsStarted | PoolingEnabled |        Mean | Allocated |
|---------- |---------- |--------------- |------------:|----------:|
| AsyncLazy |     False |          False |    29.77 ns |      80 B |
| AsyncLazy |     False |           True |    29.58 ns |      80 B |
| AsyncLazy |      True |          False | 1,388.11 ns |     464 B |
| AsyncLazy |      True |           True |   917.77 ns |      80 B |

This PR:
    
|    Method | IsStarted | PoolingEnabled |      Mean | Allocated |
|---------- |---------- |--------------- |----------:|----------:|
| AsyncLazy |     False |          False |  27.28 ns |      64 B |
| AsyncLazy |     False |           True |  24.97 ns |      64 B |
| AsyncLazy |      True |          False | 846.32 ns |     336 B |
| AsyncLazy |      True |           True | 537.03 ns |      64 B |